requirements.txt
ray
fastapi
uvicorn
ray[serve]
transformers
torch
pillow
python-multipart
datasets